Fix transport channel initialization
